// packages/frontend/src/generated/types/DispatchMeasurementResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `dispatch_storage_measurement` entity.
 */
export interface DispatchMeasurementResponse {
  /**
   * @type string,null
   */
  afterDensity?: string | null
  /**
   * @type string,null
   */
  afterHeight?: string | null
  /**
   * @type string
   */
  afterMass: string
  /**
   * @type string,null
   */
  afterVolume?: string | null
  /**
   * @type string,null
   */
  beforeDensity?: string | null
  /**
   * @type string,null
   */
  beforeHeight?: string | null
  /**
   * @type string
   */
  beforeMass: string
  /**
   * @type string,null
   */
  beforeVolume?: string | null
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string, uuid
   */
  dispatchDocId: string
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string, uuid
   */
  storageId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}